// src/cloudformation/types/cloudformation-model.ts
export interface CommonResource {
    _id?: string;
    _arn?: string;
}
// ...
export type StringType =
    | string
    | Ref
    | FnJoin
    | FnSub
    | FnGetAtt
    | FnSelect
    | FnBase64
    | FnImportValue
    | FnFindInMap
    | FnToJsonString;
// ...
export interface Ref {
    Ref: RefAttr;
}
// ...
export interface FnSub {
    'Fn::Sub': string | [string, { [key: string]: FnSubAttr }];
}
// ...
export interface FnJoin {
    'Fn::Join': [string, FnJoinAttr[]];
}
// ...
export interface CdkMetadata {
    [key: string]: unknown;
}
// ...
export interface IAMRoleResource extends CommonResource {
    Type: 'AWS::IAM::Role';
    Properties: {
        AssumeRolePolicyDocument: PolicyDocument | AnyMap;
        Description?: StringType;
        ManagedPolicyArns?: StringType[];
        MaxSessionDuration?: number;
        Path?: StringType;
        PermissionsBoundary?: StringType;
        Policies?: IAMRolePolicy[];
        RoleName?: string;
        Tags?: Tag[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}

/**
 * Represents an AWS IAM Policy resource.
 * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-iam-policy.html
 */
export interface IAMPolicyResource extends CommonResource {
    Type: 'AWS::IAM::Policy';
    Properties: {
        Groups?: StringType[];
        PolicyDocument: PolicyDocument | AnyMap;
        PolicyName: string;
        Roles?: StringType[];
        Users?: StringType[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface LambdaFunctionResource extends CommonResource {
    Type: 'AWS::Lambda::Function';
    Properties: {
        Architectures?: string[];
        Code: LambdaFunctionResourceCode;
        CodeSigningConfigArn?: StringType;
        DeadLetterConfig?: LambdaFunctionResourceDlq;
        Environment?: LambdaFunctionResourceEnv;
        EphemeralStorage?: LambdaFunctionEphemeralStorage;
        FileSystemConfigs?: LambdaFunctionFileSystemConfig[]; // Changed: this must be an array
        FunctionName?: StringType;
        Handler?: StringType;
        ImageConfig?: LambdaFunctionImageConfig;
        KmsKeyArn?: StringType;
        Layers?: StringType[];
        LoggingConfig?: LambdaFunctionLoggingConfig;
        MemorySize?: number | string;
        PackageType?: StringType; // Image | Zip
        RecursiveLoop?: StringType; // Allow | Terminate (Note: not standard in CloudFormation)
        ReservedConcurrentExecutions?: number | string;
        Role: StringType;
        Runtime?: StringType;
        RuntimeManagementConfig?: LambdaFunctionRuntimeManagementConfig;
        SnapStart?: LambdaFunctionSnapStart;
        Tags?: Tag[];
        Timeout?: number | string;
        TracingConfig?: LambdaFunctionTracingConfig;
        VpcConfig?: LambdaFunctionVpcConfig;
        [propertyName: string]: unknown;
    };
    DependsOn?: string | string[];
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}

/**
 * Represents an AWS SQS Queue resource.
 * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-sqs-queue.html
 */
export interface SQSQueueResource extends CommonResource {
    Type: 'AWS::SQS::Queue';
    Properties: {
        ContentBasedDeduplication?: boolean;
        DeduplicationScope?: StringType;
        DelaySeconds?: number;
        FifoQueue?: boolean;
        FifoThroughputLimit?: StringType;
        KmsDataKeyReusePeriodSeconds?: number;
        KmsMasterKeyId?: StringType;
        MaximumMessageSize?: number;
        MessageRetentionPeriod?: number;
        QueueName?: StringType;
        ReceiveMessageWaitTimeSeconds?: number;
        RedriveAllowPolicy?: AnyMap;
        RedrivePolicy?: AnyMap;
        SqsManagedSseEnabled?: boolean;
        Tags?: Tag[];
        VisibilityTimeout?: number;
        [propertyName: string]: unknown;
    };
    UpdateReplacePolicy?: string;
    DeletionPolicy?: string;
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface SNSTopicResource extends CommonResource {
    Type: 'AWS::SNS::Topic';
    Properties: {
        ArchivePolicy?: AnyMap;
        ContentBasedDeduplication?: boolean;
        DataProtectionPolicy?: AnyMap;
        DeliveryStatusLogging?: SNSTopicDeliveryStatusLogging[];
        DisplayName?: StringType;
        FifoThroughputScope?: StringType;
        FifoTopic?: boolean;
        KmsMasterKeyId?: StringType;
        SignatureVersion?: StringType;
        Subscription?: SNSTopicSubscription[];
        Tags?: Tag[];
        TopicName?: StringType;
        TracingConfig?: StringType;
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}

/**
 * Represents an AWS SNS Topic Policy resource.
 * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-sns-topicpolicy.html
 */
export interface SNSTopicPolicyResource extends CommonResource {
    Type: 'AWS::SNS::TopicPolicy';
    Properties: {
        PolicyDocument: PolicyDocument | AnyMap;
        Topics: StringType[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}

/**
 * Represents an AWS SNS Subscription resource.
 * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-sns-subscription.html
 */
export interface SNSSubscriptionResource extends CommonResource {
    Type: 'AWS::SNS::Subscription';
    Properties: {
        DeliveryPolicy?: AnyMap;
        Endpoint?: StringType;
        FilterPolicy?: AnyMap;
        FilterPolicyScope?: StringType;
        Protocol: StringType;
        RawMessageDelivery?: boolean;
        RedrivePolicy?: AnyMap;
        Region?: StringType;
        ReplayPolicy?: AnyMap;
        SubscriptionRoleArn?: StringType;
        TopicArn: StringType;
        [propertyName: string]: unknown;
    };
    DependsOn?: string | string[];
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface LambdaPermissionResource extends CommonResource {
    Type: 'AWS::Lambda::Permission';
    Properties: {
        Action: StringType;
        EventSourceToken?: StringType;
        FunctionName: StringType;
        FunctionUrlAuthType?: StringType;
        Principal: StringType;
        PrincipalOrgID?: StringType;
        SourceAccount?: StringType;

        SourceArn?: StringType;
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}

/**
 * Represents an AWS CloudWatch Log Group resource.
 *
 * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-logs-loggroup.html
 */
export interface LogGroupResource extends CommonResource {
    Type: 'AWS::Logs::LogGroup';
    Properties: {
        DataProtectionPolicy?: AnyMap;
        FieldIndexPolicies?: AnyMap[];
        KmsKeyId?: StringType;
        LogGroupClass?: StringType;
        LogGroupName?: StringType;
        RetentionInDays?: number;
        Tags?: Tag[];
        [propertyName: string]: unknown;
    };
    UpdateReplacePolicy?: string;
    DeletionPolicy?: string;
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface ApiGatewayDeploymentResource extends CommonResource {
    Type: 'AWS::ApiGateway::Deployment';
    Properties: {
        DeploymentCanarySettings?: ApiGatewayDeploymentResourceDeploymentCanarySettings;
        Description?: StringType;
        RestApiId: StringType;
        StageDescription?: ApiGatewayDeploymentResourceStageDescription;
        StageName?: StringType;
        [propertyName: string]: unknown;
    };
    DependsOn?: string | string[];
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface ApiGatewayRequestValidatorResource extends CommonResource {
    Type: 'AWS::ApiGateway::RequestValidator';
    Properties: {
        Name?: StringType;
        RestApiId: StringType;
        ValidateRequestBody?: boolean;
        ValidateRequestParameters?: boolean;
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface EventBusResource extends CommonResource {
    Type: 'AWS::Events::EventBus';
    Properties: {
        DeadLetterConfig?: EventBusResourceDeadLetterConfig;
        Description?: StringType;
        EventSourceName?: StringType;
        KmsKeyIdentifier?: StringType;
        Name: StringType;
        Policy?: PolicyDocument | AnyMap;
        Tags?: Tag[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface EventsRuleResource extends CommonResource {
    Type: 'AWS::Events::Rule';
    Properties: {
        Description?: StringType;
        EventBusName?: StringType;
        EventPattern?: AnyMap;
        Name?: StringType;
        RoleArn?: StringType;
        ScheduleExpression?: StringType; // e.g., cron(0 20 * * ? *)
        State?: StringType; // DISABLED | ENABLED | ENABLED_WITH_ALL_CLOUDTRAIL_MANAGEMENT_EVENTS
        Targets?: EventsRuleResourceTarget[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface AWSECSTaskDefinition extends CommonResource {
    Type: 'AWS::ECS::TaskDefinition';
    Properties: {
        ContainerDefinitions?: ContainerDefinition[];
        Cpu?: StringType;
        EnableFaultInjection?: boolean;
        EphemeralStorage?: EphemeralStorage;
        ExecutionRoleArn?: StringType;
        Family?: StringType;
        InferenceAccelerators?: InferenceAccelerator[];
        IpcMode?: StringType; // host | task | none
        Memory?: StringType;
        NetworkMode?: StringType; // bridge | host | awsvpc | none
        PidMode?: StringType; // host | task
        PlacementConstraints?: TaskDefinitionPlacementConstraint[];
        ProxyConfiguration?: ProxyConfiguration;
        RequiresCompatibilities?: StringType[];
        RuntimePlatform?: RuntimePlatform;
        Tags?: Tag[];
        TaskRoleArn?: StringType;
        Volumes?: Volume[];
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface PipesPipeResource extends CommonResource {
    Type: 'AWS::Pipes::Pipe';
    Properties: {
        Description?: StringType;
        DesiredState?: StringType;
        Enrichment?: StringType;
        EnrichmentParameters?: PipeEnrichmentParameters;
        KmsKeyIdentifier?: StringType;
        LogConfiguration?: PipeLogConfiguration;
        Name?: StringType;
        RoleArn: StringType;
        Source: StringType;
        SourceParameters?: PipeSourceParameters;
        Tags?: Tag[];
        Target: StringType;
        TargetParameters?: PipeTargetParameters;
        [propertyName: string]: unknown;
    };
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export interface StepFunctionsStateMachineResource extends CommonResource {
    Type: 'AWS::StepFunctions::StateMachine';
    Properties: {
        Definition?: AnyMap;
        DefinitionS3Location?: S3Location;
        DefinitionString?: StringType;
        DefinitionSubstitutions?: ObjectOfString;
        EncryptionConfiguration?: EncryptionConfiguration;
        LoggingConfiguration?: LoggingConfiguration;
        RoleArn?: StringType;
        StateMachineName?: StringType;
        StateMachineType?: StringType;
        Tags?: Tag[];
        TracingConfiguration?: TracingConfiguration;
        [propertyName: string]: unknown;
    };
    UpdateReplacePolicy?: string;
    DeletionPolicy?: string;
    Metadata?: CdkMetadata;

    [key: string]: unknown;
}
// ...
export type CloudFormationResource =
    | AWSECSTaskDefinition
    | EventBusResource
    | EventsRuleResource
    | IAMPolicyResource
    | IAMRoleResource
    | LambdaFunctionResource
    | LambdaPermissionResource
    | LogGroupResource
    | PipesPipeResource
    | SNSSubscriptionResource
    | SNSTopicPolicyResource
    | SNSTopicResource
    | SQSQueueResource
    | StepFunctionsStateMachineResource
    | ApiGatewayRestApiResource
    | ApiGatewayDeploymentResource
    | ApiGatewayStageResource
    | ApiGatewayResourceResource
    | ApiGatewayMethodResource
    | ApiGatewayAuthorizerResource
    | DynamoDBTableResource
    | ApiGatewayRequestValidatorResource
    | ArchiveResource
    | EventBusPolicyResource
    | EventsConnectionResource
    | EventsApiDestinationResource
    | ApiGatewayModelResource
    | CDKMetadataResource
    | ElasticLoadBalancingV2TargetGroupResource
    | AWSElasticLoadBalancingV2ListenerRule
    | EcsServiceResource
    | ApiGatewayV2StageResource
    | ApiGatewayV2AuthorizerResource
    | ApiGatewayV2ModelResource
    | ApiGatewayV2DeploymentResource
    | AwsS3BucketResource
    | GenericResource;